puzzle_backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
Pillow==10.4.0
django-import-export==4.1.1
django-cors-headers==4.4.0
python-dotenv==1.0.1
openpyxl==3.1.2
# Тестовые зависимости (убедитесь, что они есть)
pytest==7.4.3
pytest-django==4.7.0